update java swing 03

+++ JavaSwing03/src/main/java/org/example/controller/CourseController.java
@@ -3,9 +3,9 @@ package org.example.controller;
 import org.example.model.DAO.concrete.MySQLCourse;
 import org.example.model.DAO.interfaces.CourseDAO;
 import org.example.model.Course;
-import org.example.view.MainView;
 import org.example.view.Course.CourseView;
 
+import javax.swing.*;
 import javax.swing.event.ListSelectionEvent;
 import javax.swing.event.ListSelectionListener;
 import java.awt.event.ActionEvent;
@@ -14,8 +14,8 @@ import java.sql.SQLException;
 
 public class CourseController {
 
-    private CourseDAO courseDAO;
-    private CourseView courseView;
+    private final CourseDAO courseDAO;
+    private final CourseView courseView;
 
     public CourseController(CourseView courseView){
         this.courseDAO = new MySQLCourse();
@@ -66,9 +66,14 @@ public class CourseController {
                     return;
                 Course course = courseView.getCourseInfo();
                 try {
-                    courseDAO.delete(course);
-                    showCourseList();
-                    courseView.resetTextfield();
+                    int op = courseView.showConfirm("Delete course id = "+course.getId());
+                    if(op==JOptionPane.YES_OPTION){
+                        courseDAO.delete(course);
+                        showCourseList();
+                        courseView.resetTextfield();
+                    }
+                    else return;
+
                 } catch (SQLException ex) {
                     throw new RuntimeException(ex);
                 }
